Reorganizing components in folders
