+++ components/results/ResultsRetryButton.tsx
